+++ stream-chat-android-core/src/testFixtures/kotlin/io/getstream/chat/android/Mother.kt
@@ -379,7 +379,7 @@ public fun randomChannel(
     team: String = randomString(),
     hidden: Boolean? = randomBoolean(),
     hiddenMessagesBefore: Date? = randomDate(),
-    ownCapabilities: Set<String> = setOf(),
+    ownCapabilities: Set<String> = randomChannelCapabilities(),
     extraData: Map<String, Any> = emptyMap(),
     membership: Member? = randomMember(),
 ): Channel = Channel(
@@ -470,6 +470,7 @@ public fun randomConfig(
     automodBehavior: String = randomString(),
     blocklistBehavior: String = randomString(),
     commands: List<Command> = emptyList(),
+    pollsEnabled: Boolean = randomBoolean(),
 ): Config = Config(
     createdAt = createdAt,
     updatedAt = updatedAt,
@@ -491,6 +492,7 @@ public fun randomConfig(
     automodBehavior = automodBehavior,
     blocklistBehavior = blocklistBehavior,
     commands = commands,
+    pollsEnabled = pollsEnabled,
 )
 
 public fun randomChannelConfig(type: String = randomString(), config: Config = randomConfig()): ChannelConfig =