+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/viewmodel/messages/MessageListViewModelBinding.kt
@@ -112,6 +112,7 @@ public fun MessageListViewModel.bindView(
     view.setMessageUserUnblockHandler {
         onEvent(MessageListViewModel.Event.UnblockUser(it.user.id))
     }
+    view.setOnPauseAudioRecordingAttachmentsHandler { pauseAudioRecordingAttachments() }
 
     ownCapabilities.observe(lifecycleOwner) {
         view.setOwnCapabilities(it)